правка создания задачи + изменения БД + модуль для изменения профиля + начало создания модуля Портфолио

+++ database/migrations/2022_02_25_150955_field_views_number_add.php
@@ -14,7 +14,7 @@ class FieldViewsNumberAdd extends Migration
     public function up()
     {
         Schema::table('tasks', function (Blueprint $table) {
-            $table->bigInteger('views_number')->default(0);
+            $table->integer('views_number')->default(0);
         });
     }
